+++ tests/tests.rs
@@ -14,7 +14,7 @@ fn add_string_full(){
     let item_id:String = "World".to_string();
     tree.add_string(item_string.clone(), item_id);
     let sstring = tree.find(&item_string);
-    assert_eq!(sstring, vec![(&"World".to_string(), &(0 as u32))]);
+    assert_eq!(sstring, vec![(&"World".to_string(), &(0 as usize))]);
 }
 
 #[test]
@@ -26,7 +26,7 @@ fn add_string_set(){
     for (string,id) in it{
         tree.add_string(string.clone(), id.clone());
     }
-    assert_eq!(tree.find(&"XYZ".to_string().chars().collect()), vec![(&"first".to_string(), &(14 as u32))]);
+    assert_eq!(tree.find(&"XYZ".to_string().chars().collect()), vec![(&"first".to_string(), &(14 as usize))]);
 }
 
 #[test]
@@ -59,12 +59,12 @@ fn exact_pattern_match(){
     let item_id:String = "World2".to_string();
     tree.add_string(item_string.clone(), item_id);
     println!("{}", serde_json::to_string(&tree).unwrap());
-    let matches: HashSet<(&String, &u32)> = tree.find(&"XYZ".to_string().chars().collect()).into_iter().collect();
-    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(14 as u32))]));
-    let matches: HashSet<(&String, &u32)> = tree.find(&"GATTA".to_string().chars().collect()).into_iter().collect();
-    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(0 as u32)), (&"World".to_string(), &(7 as u32)), (&"World".to_string(), &(17 as u32)), (&"World".to_string(), &(24 as u32))]));
-    let matches: HashSet<(&String, &u32)> = tree.find(&"ANA".to_string().chars().collect()).into_iter().collect();
-    assert_eq!(matches, HashSet::from([(&"World2".to_string(), &(1 as u32))]));
-    let matches: HashSet<(&String, &u32)> = tree.find(&"AT".to_string().chars().collect()).into_iter().collect();
-    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(1 as u32)), (&"World".to_string(), &(8 as u32)), (&"World".to_string(), &(18 as u32)), (&"World".to_string(), &(25 as u32)), (&"World2".to_string(), &(3 as u32))]));
+    let matches: HashSet<(&String, &usize)> = tree.find(&"XYZ".to_string().chars().collect()).into_iter().collect();
+    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(14 as usize))]));
+    let matches: HashSet<(&String, &usize)> = tree.find(&"GATTA".to_string().chars().collect()).into_iter().collect();
+    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(0 as usize)), (&"World".to_string(), &(7 as usize)), (&"World".to_string(), &(17 as usize)), (&"World".to_string(), &(24 as usize))]));
+    let matches: HashSet<(&String, &usize)> = tree.find(&"ANA".to_string().chars().collect()).into_iter().collect();
+    assert_eq!(matches, HashSet::from([(&"World2".to_string(), &(1 as usize))]));
+    let matches: HashSet<(&String, &usize)> = tree.find(&"AT".to_string().chars().collect()).into_iter().collect();
+    assert_eq!(matches, HashSet::from([(&"World".to_string(), &(1 as usize)), (&"World".to_string(), &(8 as usize)), (&"World".to_string(), &(18 as usize)), (&"World".to_string(), &(25 as usize)), (&"World2".to_string(), &(3 as usize))]));
 }
